Enable to specify traffic threshold for alerts by command line
Previously, there was no way to specify on the command line upon
which traffic threshold an alert should be triggered. This was
bound to make testing of alerting slow because one would have to
generate a lot of requests manually to trigger an alert.

This commit fixes it by enabling to set this option by command line.