Update README for better command line option example
Previously, README presented a non existing option as an example
of usage. This was due to no real option existing at the time and
was both a lost occasion to give a real usage example as well as
confusing.

This commit fixes it by replacing it with a real option example.