Limit stats to 5 items per category
Previously, stats on requests with errors and section hits
were displaying items ordered in some way, but not limited. Thus
it could theoretically grow to an infinite size, including on
web interface. This defeated the purpose of wanting to resurface
only most important information.

This commit fixes it by limiting item count to 5 in each category.